requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pymongo==4.10.1
pydantic==2.10.3
langchain==0.3.15
langchain-google-genai==2.0.8
google-generativeai==0.8.3
langdetect==1.0.9
python-dotenv==1.0.1
requests==2.32.3
flask==3.1.0 